API

DataToFunctions.get_function — Function

get_function(data::AbstractArray; super_sampling=2, extrapolation_bc=Flat(), interp_type=Interpolations.BSpline(Linear()))

returns a function dat(shift, zoom) which generates a shifted and scaled version of the original data. This is useful for fitting with a function which is itself defined by measured data.

Arguments

data: The data to represent by the function dat super_sampling: The factor by which the data is internally represented as a supersampled version (Fourier-based upsampling, see FourierTools.resample) extrapolation_bc: The extrapolation boundary condition to select for values outside the range. By default the value 0.0 is used. Other options are Flat(), or Line(), See the package Interpolation for details. interp_type: The type of interpolation to use. See the package Interpolation for details.

DataToFunctions.add_dim — Function

add_dim(cind)

adds a dimension to a CartesianIndex

cind: A CartesianIndex

DataToFunctions.red_dim — Function

red_dim(svec::SVector{S,T})

removes the last dimension of a SVector to convert it from a homogeneous to a Cartesian coordinates

svec::SVector{S,T}: A SVector

DataToFunctions.red_dim_apply — Function

red_dim_apply(fct, svec::SVector{S,T})

applies a function to a SVector by removing the last dimension to convert it from a homogeneous to a Cartesian coordinates

fct: The function to apply svec::SVector{S,T}: A SVector

red_dim_apply(fct, tup::NTuple{S,T})

applies a function to a Tuple by removing the last dimension to convert it from a homogeneous to a Cartesian coordinates

fct: The function to apply tup::NTuple{S,T}: A Tuple

Missing docstring.

Missing docstring for idx_appy. Check Documenter's build log for details.

DataToFunctions.mat_mul — Function

mat_mul(t::SVector{N, T2}, matrix_c::SMatrix{N,N,T})

multiplies a SVector with a SMatrix

t::SVector{N, T2}: The SVector to multiply matrix_c::SMatrix{N,N,T}: The SMatrix to multiply with

DataToFunctions.func_transform — Function

func_transform(t, coord_transform_func::Function)::SVector

applies a coordinate transformation function to an array or CartesianIndex and returns the transformed array

t: The array or CartesianIndex to transform coord_transform_func::Function: The function to apply the transformation

Missing docstring.

Missing docstring for func_transform_tup. Check Documenter's build log for details.

Missing docstring.

Missing docstring for apply_transform. Check Documenter's build log for details.

Missing docstring.

Missing docstring for apply_transform_homogen. Check Documenter's build log for details.

Missing docstring.

Missing docstring for apply_transform_affine. Check Documenter's build log for details.

DataToFunctions.get_function_tuple — Function

get_function_tuple(data::AbstractArray, fct_tup::Function; super_sampling=2, extrapolation_bc=Flat(), interp_type=Interpolations.BSpline(Linear()))

returns a function dat(shift, zoom) which generates a shifted and scaled version of the original data. This is useful for fitting with a function which is itself defined by measured data.

Arguments

data: The data to represent by the function dat fct_tup: The function to apply to the data super_sampling: The factor by which the data is internally represented as a supersampled version (Fourier-based upsampling, see FourierTools.resample) extrapolation_bc: The extrapolation boundary condition to select for values outside the range. By default the value 0.0 is used. Other options are Flat(), or Line(), See the package Interpolation for details. interp_type: The type of interpolation to use. See the package Interpolation for details.

Example

DataToFunctions.get_function_svec — Function

get_function_svec(data::AbstractArray, fct_hom::Function; super_sampling=1, extrapolation_bc=Flat(), interp_type=Interpolations.BSpline(Linear()))

returns a function interpolated(params) which generates a transformed version of the original data parameterized by transform parameters.

Arguments

data: The data to represent by the function dat fct_hom: The function to apply to the data super_sampling: The factor by which the data is internally represented as a supersampled version (Fourier-based upsampling, see FourierTools.resample) extrapolation_bc: The extrapolation boundary condition to select for values outside the range. By default the value 0.0 is used. Other options are Flat(), or Line(), See the package Interpolation for details. interp_type: The type of interpolation to use. See the package Interpolation for details.

DataToFunctions.get_function_affine — Function

get_function_affine(data::AbstractArray; super_sampling=1, extrapolation_bc=Flat(), interp_type=Interpolations.BSpline(Linear()))

returns a function interpolated() which generates a transformed version of the original data parameterized by transform parameters or by transformation matrix. This is useful for fitting with a function which is itself defined by measured data. The returned function supports two ways to be used, with an affine transform matrix matrix_c as in input or with a vector p of parameters.

Arguments

data: The data to represent by the function dat super_sampling: The factor by which the data is internally represented as a supersampled version (Fourier-based upsampling, see FourierTools.resample) extrapolation_bc: The extrapolation boundary condition to select for values outside the range. By default the value 0.0 is used. Other options are Flat(), or Line(), See the package Interpolation for details. interp_type: The type of interpolation to use. See the package Interpolation for details.

DataToFunctions.get_function_poly — Function

get_function_poly(data::AbstractArray, order; super_sampling=1, extrapolation_bc=Flat(), interp_type=Interpolations.BSpline(Linear()))

returns a function interpolated(p, [out]) which generates a transformed version of the original data parameterized by transform parameters. This is useful for fitting with a function which is itself defined by measured data. The returned function supports two ways to be used, with an affine transform matrix p as in input or with a vector p of parameters. The optional argument out can be used to store the result of the transformation.

Arguments

data: The data to represent by the function dat order: The order of the polynomial to use for the transformation super_sampling: The factor by which the data is internally represented as a supersampled version (Fourier-based upsampling, see FourierTools.resample) extrapolation_bc: The extrapolation boundary condition to select for values outside the range. By default the value 0.0 is used. Other options are Flat(), or Line(), See the package Interpolation for details. interp_type: The type of interpolation to use. See the package Interpolation for details.

DataToFunctions.split_tuple — Function

split_tuple(t::NTuple{S,T},::Val{numvars}) where {S,T,numvars}

Split a tuple into numvars parts packed into a tuple of tuples. The tuple t is assumed to have a length that is a multiple of numvars.

Example:

julia> t = (1,2,3,4,5,6)
julia> split_tuple(t, Val{2}) 
((1,2), (3,4), (5,6))